admiralpeds: Pediatrics Extension Package for ADaM in 'R' Asset Library

A toolbox for programming Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, https://www.cdisc.org/standards/foundational/adam). The package is an extension package of the 'admiral' package for pediatric clinical trials.

See also

Author

Maintainer: Fanny Gautier fanny.gautier@cytel.com

Authors:

Ross Farrugia

Zelos Zhu

Sukalpo Saha

Lina Patil

Samia Kabi

Other contributors:

Laura Liao [contributor]

Remigiusz Kudlacz [contributor]

Pierre Wallet [contributor]

Amin Sherzad [contributor]

David Freedman [contributor]

Mahmoud Hamza [contributor]

F. Hoffmann-La Roche AG [copyright holder, funder]

Cytel [copyright holder, funder]

Pfizer Inc. [copyright holder, funder]

Atorus Research [copyright holder, funder]